requirements.txt
jupyter
numpy
scipy
pandas
nose
mako
cython
pyzmq
matplotlib
geopy
matplotlib-venn
pypng
sphinx
numpydoc
requests
pyyaml
pytest-cov
coverage
